requirements.txt
-e .
aws_cdk.core>=1.80
aws_cdk.aws_s3>=1.80
aws_cdk.aws_lambda>=1.80
aws_cdk.aws_logs>=1.80
aws_cdk.aws_s3_notifications>=1.80
aws_cdk.aws_dynamodb>=1.80
aws_cdk.aws_ecr>=1.80
aws_cdk.aws_ec2>=1.80
aws_cdk.aws_autoscaling>=1.80
aws_cdk.aws_codecommit>=1.80
aws_cdk.aws_codebuild>=1.80
aws_cdk.aws_codedeploy>=1.80
aws_cdk.aws_codepipeline>=1.80
aws_cdk.aws_codepipeline_actions>=1.80
aws_cdk.aws_ecs>=1.80
aws_cdk.aws_ecs_patterns>=1.80